Various updates
Implements mil time for arithmetic operations, tries to begin assessing both SLA101 and 191 specific rules, modularizes the SLA rules to make referencing easier, and attempts to assess the consecutive activities for a single facilitator. Bugs are still prevalent all throughout the program, such as assessing ALL courses under the consecutive rule as of the current iteration.