Introduces light unit testing & reformats pop organization
Unit testing is intended to serve as a highly direct method to assess the accuracy of my `fitness.py` file. Currently, I am planning on using the `population.py` file to create *one* "test schedule" with a known fitness value, and hard code this into the unit test for validating the completeness of the `fitness.py` file.